+++ CHANGELOG.md
@@ -3,9 +3,7 @@
 ## [Unreleased]
 
 - Added `ESActionBlockContainer` class.
-- Added UIBarButtonItem `actionBlock` property.
-- Added UIGestureRecognizer action block methods: `-initWithActionBlock:`, `-addActionBlock:`, `removeAllActionBlocks`.
-- Added UIControl action block methods.
+- Added action block methods for `UIBarButtonItem`, `UIGestureRecognizer`, `UIControl`.
 
 ## 3.1.0 (2019-05-23)